+++ CHANGELOG.markdown
@@ -4,6 +4,7 @@
 * SMT
   * Migrate toysmt to use newly developed [language-smtlib](https://hackage.haskell.org/package/language-smtlib) from `Smtlib` fork (#206)
   * Update the semantics of zero division on bitvectors to conform to SMT-LIB >=2.6 (#209)
+  * `toysmt` now suppresses the prompt and does not use haskeline when stdin is not a terminal
 * Converter and `toyconvert` command
   * Do not produce `obj<T` constraint in `wbo2pb` when all interpretations are admissible (#157)
   * Optimize `unconstrainPB` a little (#159, #161)